[FIXED JENKINS-37005] Warn users using properties outside multibranch
This doesn't change actual behavior. It simply logs to the console
when the properties step is invoked outside of a multibranch job,
including listing any properties that are removed. The goal is to
decrease confusion if/when a non-multibranch job suddenly loses its
triggers, etc due to a properties step being executed in the job.